Setup rndc

+++ scripts/register.sh
@@ -10,7 +10,6 @@ fi
 . "${PROJECT_PATH}/scripts/default.conf"
 
 CBSD_WORKDIR=`sysrc -n cbsd_workdir`
-TEMP_FILE=`mktemp ${CBSD_WORKDIR}/jails-data/resolver-data/tmp/tmp.XXXXXX`
 JAIL_NAME=${jname}
 JAIL_IP=`echo ${ip4_addr} | cut -f 1 -d '/'`
 ACTION="${1}"
@@ -22,14 +21,17 @@ if [ "${ACTION}" = "deregister" ]; then
     PF_ACTION="delete"
 fi
 
-sed \
-    -e "s/JAIL_NAME/${JAIL_NAME}/g" \
-    -e "s/JAIL_IP/${JAIL_IP}/g" \
-    -e "s/RESOLVER_IP/${RESOLVER_IP}/g" \
-    ${TEMPLATE} \
-    >${TEMP_FILE}
+if [ "${JAIL_NAME}" != "resolver" ]; then
+  TEMP_FILE=`mktemp ${CBSD_WORKDIR}/jails-data/resolver-data/tmp/tmp.XXXXXX`
+  sed \
+      -e "s/JAIL_NAME/${JAIL_NAME}/g" \
+      -e "s/JAIL_IP/${JAIL_IP}/g" \
+      -e "s/RESOLVER_IP/${RESOLVER_IP}/g" \
+      ${TEMPLATE} \
+      >${TEMP_FILE}
 
-cbsd jexec jname=resolver nsupdate -k /usr/local/etc/namedb/cbsd.key ${TEMP_FILE#${CBSD_WORKDIR}/jails-data/resolver-data}
+  cbsd jexec jname=resolver nsupdate -k /usr/local/etc/namedb/cbsd.key ${TEMP_FILE#${CBSD_WORKDIR}/jails-data/resolver-data}
+  rm -rf ${TEMP_FILE}
+fi
 pfctl -t cbsd -T ${PF_ACTION} ${JAIL_IP}
-rm -rf ${TEMP_FILE}
